requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
typing_extensions==4.14.0
Werkzeug==3.1.3
